+++ src/OpenTelemetry/Metrics/CompositeMetricReader.cs
@@ -75,11 +75,13 @@ namespace OpenTelemetry.Metrics
         protected override bool OnCollect(int timeoutMilliseconds = Timeout.Infinite)
         {
             var result = true;
-            var sw = Stopwatch.StartNew();
+            var sw = timeoutMilliseconds == Timeout.Infinite
+                ? null
+                : Stopwatch.StartNew();
 
             for (var cur = this.head; cur != null; cur = cur.Next)
             {
-                if (timeoutMilliseconds == Timeout.Infinite)
+                if (sw == null)
                 {
                     result = cur.Value.Collect(Timeout.Infinite) && result;
                 }
@@ -99,11 +101,13 @@ namespace OpenTelemetry.Metrics
         protected override bool OnShutdown(int timeoutMilliseconds)
         {
             var result = true;
-            var sw = Stopwatch.StartNew();
+            var sw = timeoutMilliseconds == Timeout.Infinite
+                ? null
+                : Stopwatch.StartNew();
 
             for (var cur = this.head; cur != null; cur = cur.Next)
             {
-                if (timeoutMilliseconds == Timeout.Infinite)
+                if (sw == null)
                 {
                     result = cur.Value.Shutdown(Timeout.Infinite) && result;
                 }