+++ src/OpenTelemetry/Metrics/MeterProviderSdk.cs
@@ -30,6 +30,7 @@ namespace OpenTelemetry.Metrics
         internal const int MaxMetrics = 1000;
         internal int ShutdownCount;
         private readonly Metric[] metrics;
+        private readonly Metric[] metricsCurrentBatch;
         private readonly List<object> instrumentations = new List<object>();
         private readonly List<Func<Instrument, MetricStreamConfiguration>> viewConfigs;
         private readonly object collectLock = new object();
@@ -50,6 +51,7 @@ namespace OpenTelemetry.Metrics
             this.Resource = resource;
             this.viewConfigs = viewConfigs;
             this.metrics = new Metric[MaxMetrics];
+            this.metricsCurrentBatch = new Metric[MaxMetrics];
 
             AggregationTemporality temporality = AggregationTemporality.Cumulative;
 
@@ -135,7 +137,7 @@ namespace OpenTelemetry.Metrics
                         // which will apply defaults.
                         // Users can turn off this default
                         // by adding a view like below as the last view.
-                        // .AddView(instrumentName: "*", new MetricStreamConfiguration() { Aggregation = Aggregation.Drop })
+                        // .AddView(instrumentName: "*", MetricStreamConfiguration.Drop)
                         metricStreamConfigs.Add(null);
                     }
 
@@ -151,7 +153,21 @@ namespace OpenTelemetry.Metrics
                         for (int i = 0; i < maxCountMetricsToBeCreated; i++)
                         {
                             var metricStreamConfig = metricStreamConfigs[i];
-                            var metricStreamName = metricStreamConfig?.Name ?? instrument.Name;
+                            var meterName = instrument.Meter.Name;
+                            var metricName = metricStreamConfig?.Name ?? instrument.Name;
+                            var metricStreamName = $"{meterName}.{metricName}";
+
+                            if (!MeterProviderBuilderSdk.IsValidInstrumentName(metricName))
+                            {
+                                OpenTelemetrySdkEventSource.Log.MetricInstrumentIgnored(
+                                    metricName,
+                                    instrument.Meter.Name,
+                                    "Metric name is invalid.",
+                                    "The name must comply with the OpenTelemetry specification.");
+
+                                continue;
+                            }
+
                             if (this.metricStreamNames.Contains(metricStreamName))
                             {
                                 // TODO: Log that instrument is ignored
@@ -181,7 +197,7 @@ namespace OpenTelemetry.Metrics
                                 string[] tagKeysInteresting = metricStreamConfig?.TagKeys;
                                 double[] histogramBucketBounds = (metricStreamConfig is HistogramConfiguration histogramConfig
                                     && histogramConfig.BucketBounds != null) ? histogramConfig.BucketBounds : null;
-                                metric = new Metric(instrument, temporality, metricStreamName, metricDescription, histogramBucketBounds, tagKeysInteresting);
+                                metric = new Metric(instrument, temporality, metricName, metricDescription, histogramBucketBounds, tagKeysInteresting);
 
                                 this.metrics[index] = metric;
                                 metrics.Add(metric);
@@ -205,6 +221,8 @@ namespace OpenTelemetry.Metrics
                 this.listener.SetMeasurementEventCallback<int>((instrument, value, tags, state) => this.MeasurementRecordedLong(instrument, value, tags, state));
                 this.listener.SetMeasurementEventCallback<short>((instrument, value, tags, state) => this.MeasurementRecordedLong(instrument, value, tags, state));
                 this.listener.SetMeasurementEventCallback<byte>((instrument, value, tags, state) => this.MeasurementRecordedLong(instrument, value, tags, state));
+
+                this.listener.MeasurementsCompleted = (instrument, state) => this.MeasurementsCompleted(instrument, state);
             }
             else
             {
@@ -218,11 +236,24 @@ namespace OpenTelemetry.Metrics
 
                     try
                     {
+                        if (!MeterProviderBuilderSdk.IsValidInstrumentName(instrument.Name))
+                        {
+                            OpenTelemetrySdkEventSource.Log.MetricInstrumentIgnored(
+                                instrument.Name,
+                                instrument.Meter.Name,
+                                "Instrument name is invalid.",
+                                "The name must comply with the OpenTelemetry specification");
+
+                            return;
+                        }
+
+                        var meterName = instrument.Meter.Name;
                         var metricName = instrument.Name;
+                        var metricStreamName = $"{meterName}.{metricName}";
                         Metric metric = null;
                         lock (this.instrumentCreationLock)
                         {
-                            if (this.metricStreamNames.Contains(metricName))
+                            if (this.metricStreamNames.Contains(metricStreamName))
                             {
                                 OpenTelemetrySdkEventSource.Log.MetricInstrumentIgnored(metricName, instrument.Meter.Name, "Metric name conflicting with existing name.", "Either change the name of the instrument or change name using View.");
                                 return;
@@ -238,7 +269,7 @@ namespace OpenTelemetry.Metrics
                             {
                                 metric = new Metric(instrument, temporality, metricName, instrument.Description);
                                 this.metrics[index] = metric;
-                                this.metricStreamNames.Add(metricName);
+                                this.metricStreamNames.Add(metricStreamName);
                             }
                         }
 
@@ -259,9 +290,10 @@ namespace OpenTelemetry.Metrics
                 this.listener.SetMeasurementEventCallback<int>((instrument, value, tags, state) => this.MeasurementRecordedLongSingleStream(instrument, value, tags, state));
                 this.listener.SetMeasurementEventCallback<short>((instrument, value, tags, state) => this.MeasurementRecordedLongSingleStream(instrument, value, tags, state));
                 this.listener.SetMeasurementEventCallback<byte>((instrument, value, tags, state) => this.MeasurementRecordedLongSingleStream(instrument, value, tags, state));
+
+                this.listener.MeasurementsCompleted = (instrument, state) => this.MeasurementsCompletedSingleStream(instrument, state);
             }
 
-            this.listener.MeasurementsCompleted = (instrument, state) => this.MeasurementsCompleted(instrument, state);
             this.listener.Start();
 
             static Regex GetWildcardRegex(IEnumerable<string> collection)
@@ -277,9 +309,31 @@ namespace OpenTelemetry.Metrics
 
         internal MetricReader Reader => this.reader;
 
+        internal void MeasurementsCompletedSingleStream(Instrument instrument, object state)
+        {
+            var metric = state as Metric;
+            if (metric == null)
+            {
+                // TODO: log
+                return;
+            }
+
+            metric.InstrumentDisposed = true;
+        }
+
         internal void MeasurementsCompleted(Instrument instrument, object state)
         {
-            Console.WriteLine($"Instrument {instrument.Meter.Name}:{instrument.Name} completed.");
+            var metrics = state as List<Metric>;
+            if (metrics == null)
+            {
+                // TODO: log
+                return;
+            }
+
+            foreach (var metric in metrics)
+            {
+                metric.InstrumentDisposed = true;
+            }
         }
 
         internal void MeasurementRecordedDouble(Instrument instrument, double value, ReadOnlySpan<KeyValuePair<string, object>> tagsRos, object state)
@@ -387,14 +441,33 @@ namespace OpenTelemetry.Metrics
                         OpenTelemetrySdkEventSource.Log.MetricObserverCallbackException(exception);
                     }
 
-                    var indexSnapShot = Math.Min(this.metricIndex, MaxMetrics - 1);
-                    var target = indexSnapShot + 1;
+                    var indexSnapshot = Math.Min(this.metricIndex, MaxMetrics - 1);
+                    var target = indexSnapshot + 1;
+                    int metricCountCurrentBatch = 0;
                     for (int i = 0; i < target; i++)
                     {
-                        this.metrics[i].SnapShot();
+                        var metric = this.metrics[i];
+                        int metricPointSize = 0;
+                        if (metric != null)
+                        {
+                            if (metric.InstrumentDisposed)
+                            {
+                                metricPointSize = metric.Snapshot();
+                                this.metrics[i] = null;
+                            }
+                            else
+                            {
+                                metricPointSize = metric.Snapshot();
+                            }
+
+                            if (metricPointSize > 0)
+                            {
+                                this.metricsCurrentBatch[metricCountCurrentBatch++] = metric;
+                            }
+                        }
                     }
 
-                    return (target > 0) ? new Batch<Metric>(this.metrics, target) : default;
+                    return (metricCountCurrentBatch > 0) ? new Batch<Metric>(this.metricsCurrentBatch, metricCountCurrentBatch) : default;
                 }
                 catch (Exception)
                 {