+++ test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpMetricsExporterTests.cs
@@ -20,7 +20,6 @@ using System.Diagnostics.Metrics;
 using System.Linq;
 using System.Threading;
 using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
-using OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
@@ -56,18 +55,15 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 new KeyValuePair<string, object>("key2", "value2"),
             };
 
-            var metricReader = new BaseExportingMetricReader(new TestExporter<Metric>(RunTest))
-            {
-                PreferredAggregationTemporality = AggregationTemporality.Delta,
-            };
+            using var meter = new Meter($"{Utils.GetCurrentMethodName()}.{includeServiceNameInResource}", "0.0.1");
+
+            var exportedItems = new List<Metric>();
 
             using var provider = Sdk.CreateMeterProviderBuilder()
                 .SetResourceBuilder(resourceBuilder)
-                .AddMeter("TestMeter")
-                .AddReader(metricReader)
+                .AddMeter(meter.Name)
+                .AddReader(new BaseExportingMetricReader(new InMemoryExporter<Metric>(exportedItems)) { PreferredAggregationTemporality = AggregationTemporality.Delta })
                 .Build();
-
-            using var meter = new Meter("TestMeter", "0.0.1");
 
             var counter = meter.CreateCounter<int>("counter");
 
@@ -75,8 +71,10 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
 
             var testCompleted = false;
 
-            // Invokes the TestExporter which will invoke RunTest
-            metricReader.Collect();
+            provider.ForceFlush();
+
+            var batch = new Batch<Metric>(exportedItems.ToArray(), exportedItems.Count);
+            RunTest(batch);
 
             Assert.True(testCompleted);
 
@@ -102,7 +100,7 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests
                 Assert.Single(resourceMetric.InstrumentationLibraryMetrics);
                 var instrumentationLibraryMetrics = resourceMetric.InstrumentationLibraryMetrics.First();
                 Assert.Equal(string.Empty, instrumentationLibraryMetrics.SchemaUrl);
-                Assert.Equal("TestMeter", instrumentationLibraryMetrics.InstrumentationLibrary.Name);
+                Assert.Equal(meter.Name, instrumentationLibraryMetrics.InstrumentationLibrary.Name);
                 Assert.Equal("0.0.1", instrumentationLibraryMetrics.InstrumentationLibrary.Version);
 
                 Assert.Single(instrumentationLibraryMetrics.Metrics);