+++ test/OpenTelemetry.Tests/Metrics/MemoryEfficiencyTests.cs
@@ -24,12 +24,12 @@ namespace OpenTelemetry.Metrics.Tests
 {
     public class MemoryEfficiencyTests
     {
-        [Theory(Skip = "To be run after https://github.com/open-telemetry/opentelemetry-dotnet/issues/2524 is fixed")]
+        [Theory]
         [InlineData(AggregationTemporality.Cumulative)]
         [InlineData(AggregationTemporality.Delta)]
         public void ExportOnlyWhenPointChanged(AggregationTemporality temporality)
         {
-            using var meter = new Meter(Utils.GetCurrentMethodName(), "1.0");
+            using var meter = new Meter($"{Utils.GetCurrentMethodName()}.{temporality}");
 
             var exportedItems = new List<Metric>();
 
@@ -50,7 +50,14 @@ namespace OpenTelemetry.Metrics.Tests
 
             exportedItems.Clear();
             meterProvider.ForceFlush();
-            Assert.Empty(exportedItems);
+            if (temporality == AggregationTemporality.Cumulative)
+            {
+                Assert.Single(exportedItems);
+            }
+            else
+            {
+                Assert.Empty(exportedItems);
+            }
         }
     }
 }